+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+[Unreleased] - 2017-12-07
+-------------------------
+- Adapt to latest version of rc_dynamics, i.e., add start_slam, and stop_slam functionality
+
 [0.4.0] - 2017-09-27
 --------------------